finish refactor on District
