+++ frontend/src/js/Districts/DistrictDetails.jsx
