+++ frontend/src/js/Parties/PoliticalPartyDetails.js
@@ -4,7 +4,7 @@ import {Link} from 'react-router-dom'
 import {GridList} from 'material-ui/GridList'
 import {Timeline} from 'react-twitter-widgets'
 import RepresentativeInstance from '../Representatives/RepresentativeInstance'
-import DistrictInstance from '../Districts/DistrictInstance.js'
+import DistrictInstance from '../Districts/DistrictInstance.jsx'
 import {RingLoader} from 'react-spinners'
 import axios from 'axios'
 import Slider from 'react-slick'
