+++ frontend/src/js/Router.js
@@ -5,8 +5,8 @@ import {Route, Switch} from 'react-router-dom'
 
 import About from './About/About.jsx'
 import Alma from './Representatives/Alma.js'
-import Districts from './Districts/Districts'
-import DistrictsDetails from './Districts/DistrictDetails'
+import Districts from './Districts/Districts.jsx'
+import DistrictsDetails from './Districts/DistrictDetails.jsx'
 import PoliticalPartyDetails from './Parties/PoliticalPartyDetails'
 import Representatives from './Representatives/Representatives'
 import RepresentativeDetails from './Representatives/RepresentativeDetails'
