+++ frontend/src/js/Search.js
@@ -8,7 +8,7 @@ import ReactPaginate from 'react-paginate'
 
 import RepresentativeSingleInstance from './Representatives/RepresentativeSingleInstance'
 import PoliticalPartySingleInstance from './Parties/PoliticalPartySingleInstance'
-import DistrictInstance from './Districts/DistrictInstance'
+import DistrictInstance from './Districts/DistrictInstance.jsx'
 import axios from 'axios'
 import '../assets/css/Search.css'
 import url from '../assets/resource.json'
